+++ CHANGELOG.md
@@ -9,6 +9,7 @@ CHANGELOG
   - Make method discovery deterministic
   - Improvements to discovering operation endpoints
 - Relax heuristics on setting update mask in metadata [#515](https://github.com/pulumi/pulumi-google-native/pull/515)
+- Set force-new for storage bucket resource [#516](https://github.com/pulumi/pulumi-google-native/pull/516)
 
 ## 0.19.1 (2022-05-24)